Changed BufferLayer's canReceiveInput to check if its hidden by policy
WM does not consider window size or alpha value when setting it as the focused window. However, SF
will check if the window has a buffer and that its alpha is not 0. Because of this difference,
Input Dispatcher will not be able to send input to the window while WM thinks the window is
focused. This will cause apps to stop responding.

While we define what the intended behavior should be, this fix reverts
to previous behavior in P.

Bug: 139494112
Test: adb shell monkey 10000; make sure monkey does not get stuck
Test: check test app from can receive input b/140478820
Merged-In: I4160b49161dd1780713980707a5c911034f414b5
Change-Id: I4160b49161dd1780713980707a5c911034f414b5